+++ route/src/main/java/sparta/rooibos/route/application/dto/request/CreateRouteRequest.java
@@ -1,4 +1,4 @@
-package sparta.rooibos.route.presentation.dto.request;
+package sparta.rooibos.route.application.dto.request;
 
 import java.util.UUID;
