+++ settings.gradle
@@ -3,4 +3,4 @@ include 'common'
 include 'client'
 include 'euraka'
 // include 'product'
- include 'route'
+include 'route'
